backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
biopython==1.83
supabase==1.0.3
httpx==0.23.3
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.9
